+++ packages/moonstone/internal/SliderDecorator/SliderDecorator.js
@@ -20,7 +20,8 @@ import {validateRange} from '../validators';
 import {
 	computeProportionProgress,
 	computeBarTransform,
-	computeKnobTransform
+	computeKnobTransform,
+	parseNumber
 } from './util';
 
 /**
@@ -87,6 +88,14 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * @public
 			 */
 			backgroundProgress: PropTypes.number,
+
+			/**
+			 * Default value applied at construction when the value prop is `undefined` or `null`.
+			 *
+			 * @type {Number}
+			 * @public
+			 */
+			defaultValue: PropTypes.number,
 
 			/**
 			 * The slider can change its behavior to have the knob follow the cursor as it moves
@@ -198,7 +207,6 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			max: 100,
 			min: 0,
 			step: 1,
-			value: 0,
 			vertical: false
 		};
 
@@ -210,16 +218,24 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			this.normalizeBounds(props);
 			this.detachedKnobPosition = 0;
 
-			const
-				value = this.clamp(props.value),
-				valueText = props['aria-valuetext'] != null ? props['aria-valuetext'] : value;
+			let value, controlled = false;
+
+			if (props.value != null) {
+				controlled = true;
+				value = props.value;
+			} else {
+				value = props.defaultValue || 0;
+			}
+
+			const valueText = props['aria-valuetext'] != null ? props['aria-valuetext'] : value;
 
 			this.state = {
 				active: false,
+				controlled,
 				knobAfterMidpoint: false,
 				focused: false,
-				value: value,
-				valueText: valueText
+				value: this.clamp(value),
+				valueText
 			};
 
 			if (__DEV__) {
@@ -244,7 +260,7 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 			if ((min !== this.props.min) || (max !== this.props.max) || (value !== this.state.value) || (ariaValueText !== this.state.valueText)) {
 				this.normalizeBounds(nextProps);
-				const clampedValue = this.clamp(value);
+				const clampedValue = this.clamp(this.state.controlled ? value : this.state.value);
 				const valueText = ariaValueText != null ? ariaValueText : clampedValue;
 				this.setState({
 					value: clampedValue,
@@ -282,11 +298,21 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		updateValueJob = new Job((value) => {
 			const valueText = this.props['aria-valuetext'] != null ? this.props['aria-valuetext'] : value;
 			this.inputNode.value = value;
+			if (this.state.controlled) {
+				// FIXME: Temporarily store value for mouse down to properly fire `onChange` for controlled slider.
+				// The value will not change as componentWillReceiveProps sets state again with props.value, however
+				// the "changed" value needs to be fired on mouse up.
+				this.changedControlledValue = value;
+			}
+
 			this.setState({
 				value,
 				valueText
 			});
-			forwardChange({value}, this.props);
+			if (this.willChange) {
+				forwardChange({value}, this.props);
+				this.willChange = false;
+			}
 		}, config.changeDelay)
 
 		throttleUpdateValue = (value) => {
@@ -386,8 +412,9 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			if (this.props.disabled) return;
 
 			ev.preventDefault();
-			const parseFn = (ev.target.value % 1 !== 0) ? parseFloat : parseInt,
-				value = parseFn(ev.target.value);
+			ev.stopPropagation(); // we don't want input's onChange synthetic event to propagate
+
+			const value = parseNumber(ev.target.value);
 			this.throttleUpdateValue(value);
 
 			this.knobPosition = null;
@@ -396,19 +423,26 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		handleMouseEnter = (ev) => {
 			forwardMouseEnter(ev, this.props);
 
-			// We don't want to run this code if any mouse button is being held down. That indicates dragging.
-			if (!this.props.detachedKnob || this.props.disabled || ev.buttons || this.props.vertical) return;
+			if (!this.props.detachedKnob || this.props.disabled || this.props.vertical) return;
 
 			this.moveKnobByPointer(ev.clientX);
 		}
 
 		handleMouseMove = (ev) => {
 			forwardMouseMove(ev, this.props);
+			this.willChange = Boolean(ev.buttons);
 
 			// We don't want to run this code if any mouse button is being held down. That indicates dragging.
 			if (!this.props.detachedKnob || this.props.disabled || ev.buttons || this.props.vertical) return;
 
 			this.moveKnobByPointer(ev.clientX);
+		}
+
+		handleMouseUp = (ev) => {
+			if (ev.target.nodeName === 'INPUT') {
+				const value = parseNumber(this.state.controlled ? this.changedControlledValue : this.state.value);
+				forwardChange({value}, this.props);
+			}
 		}
 
 		handleMouseLeave = (ev) => {
@@ -435,6 +469,7 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 			if (detachedKnob) {
 				if (this.current5WayValue !== null) {
+					this.willChange = true;
 					this.throttleUpdateValue(this.clamp(this.current5WayValue));
 
 					// only clear knobPosition when not in
@@ -487,6 +522,7 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			if (detachedKnob) {
 				this.moveKnobByAmount(amount);
 			} else {
+				this.willChange = true;
 				this.throttleUpdateValueByAmount(amount);
 			}
 		}
@@ -499,6 +535,7 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			if (detachedKnob) {
 				this.moveKnobByAmount(-amount);
 			} else {
+				this.willChange = true;
 				this.throttleUpdateValueByAmount(-amount);
 			}
 		}
@@ -527,9 +564,9 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 					active={this.state.active}
 					aria-disabled={this.props.disabled}
 					aria-valuetext={this.state.valueText}
-					knobAfterMidpoint={this.state.knobAfterMidpoint}
 					focused={this.state.focused}
 					inputRef={this.getInputNode}
+					knobAfterMidpoint={this.state.knobAfterMidpoint}
 					onActivate={this.handleActivate}
 					onBlur={this.handleBlur}
 					onChange={this.handleChange}
@@ -540,6 +577,7 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 					onMouseEnter={this.handleMouseEnter}
 					onMouseLeave={this.handleMouseLeave}
 					onMouseMove={this.handleMouseMove}
+					onMouseUp={this.handleMouseUp}
 					scrubbing={(this.knobPosition != null)}
 					sliderBarRef={this.getSliderBarNode}
 					sliderRef={this.getSliderNode}