+++ packages/moonstone/CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 - `moonstone/VideoPlayer` properties `seekDisabled` and `onSeekFailed` to disable seek function
 
+### Changed
+
+- `moonstone/Slider` to fire `onChange` on mouse up and key up
+
 ### Fixed
 
 - `moonstone/ContextualPopup` padding to not overlap close button